+++ packages/atomic/src/components/search/atomic-facet-manager/atomic-facet-manager.stories.tsx
@@ -32,5 +32,5 @@ const {defaultModuleExport, exportedStory} = defaultStory(
   }
 );
 
-export default defaultModuleExport;
+export default {...defaultModuleExport, title: 'Atomic/FacetManager'};
 export const DefaultFacetManager = exportedStory;