+++ packages/atomic/src/components/search/atomic-query-error/atomic-query-error.stories.tsx
@@ -11,5 +11,5 @@ const {defaultModuleExport, exportedStory} = defaultStory(
   }
 );
 
-export default defaultModuleExport;
+export default {...defaultModuleExport, title: 'Atomic/QueryError'};
 export const DefaultQueryError = exportedStory;