+++ packages/atomic/src/components/search/facets/atomic-color-facet/atomic-color-facet.stories.tsx
@@ -95,5 +95,5 @@ const {defaultModuleExport, exportedStory} = defaultStory(
   }
 );
 
-export default defaultModuleExport;
+export default {...defaultModuleExport, title: 'Atomic/ColorFacet'};
 export const DefaultColorFacet = exportedStory;