+++ packages/headless/src/controllers/case-field/headless-case-field.test.ts
@@ -1,8 +1,8 @@
+import {configuration} from '../../app/common-reducers';
 import {
   caseAssistConfiguration,
   caseField,
   caseInput,
-  configuration,
   documentSuggestion,
 } from '../../app/reducers';
 import {