+++ packages/headless/src/controllers/core/context/headless-core-context.ts
@@ -1,10 +1,10 @@
 import {CoreEngine} from '../../../app/engine';
-import {context} from '../../../app/reducers';
 import {
   setContext,
   addContext,
   removeContext,
 } from '../../../features/context/context-actions';
+import {contextReducer as context} from '../../../features/context/context-slice';
 import {
   ContextPayload,
   ContextValue,
@@ -19,7 +19,8 @@ import {
 export type {ContextPayload, ContextValue};
 
 /**
- * The `Context` controller injects [custom contextual information](https://docs.coveo.com/en/399/) into the search requests and usage analytics search events sent from a search interface.
+ * The `Context` controller injects [custom contextual information](https://docs.coveo.com/en/3389/)
+ * into the search requests and usage analytics search events sent from a search interface.
  */
 export interface Context extends Controller {
   /**