+++ packages/headless/src/controllers/facets/range-facet/date-facet/headless-date-filter.test.ts
@@ -1,4 +1,5 @@
-import {configuration, dateFacetSet, search} from '../../../../app/reducers';
+import {configuration} from '../../../../app/common-reducers';
+import {dateFacetSet, search} from '../../../../app/reducers';
 import {updateFacetOptions} from '../../../../features/facet-options/facet-options-actions';
 import {
   registerDateFacet,