+++ packages/atomic-hosted-page/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.9](https://github.com/coveo/ui-kit/compare/@coveo/atomic-hosted-page@0.3.8...@coveo/atomic-hosted-page@0.3.9) (2023-05-11)
+
+**Note:** Version bump only for package @coveo/atomic-hosted-page
+
+## [0.3.8](https://github.com/coveo/ui-kit/compare/@coveo/atomic-hosted-page@0.3.7...@coveo/atomic-hosted-page@0.3.8) (2023-05-11)
+
+**Note:** Version bump only for package @coveo/atomic-hosted-page
+
+## [0.3.7](https://github.com/coveo/ui-kit/compare/@coveo/atomic-hosted-page@0.3.6...@coveo/atomic-hosted-page@0.3.7) (2023-05-09)
+
+**Note:** Version bump only for package @coveo/atomic-hosted-page
+
+## [0.3.6](https://github.com/coveo/ui-kit/compare/@coveo/atomic-hosted-page@0.3.5...@coveo/atomic-hosted-page@0.3.6) (2023-05-08)
+
+**Note:** Version bump only for package @coveo/atomic-hosted-page
+
+## [0.3.5](https://github.com/coveo/ui-kit/compare/@coveo/atomic-hosted-page@0.3.4...@coveo/atomic-hosted-page@0.3.5) (2023-05-04)
+
+**Note:** Version bump only for package @coveo/atomic-hosted-page
+
+## [0.3.4](https://github.com/coveo/ui-kit/compare/@coveo/atomic-hosted-page@0.3.3...@coveo/atomic-hosted-page@0.3.4) (2023-04-27)
+
+**Note:** Version bump only for package @coveo/atomic-hosted-page
+
+## [0.3.3](https://github.com/coveo/ui-kit/compare/@coveo/atomic-hosted-page@0.3.2...@coveo/atomic-hosted-page@0.3.3) (2023-04-27)
+
+**Note:** Version bump only for package @coveo/atomic-hosted-page
+
+## [0.3.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic-hosted-page@0.3.1...@coveo/atomic-hosted-page@0.3.2) (2023-04-27)
+
+**Note:** Version bump only for package @coveo/atomic-hosted-page
+
+## [0.3.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic-hosted-page@0.3.0...@coveo/atomic-hosted-page@0.3.1) (2023-04-26)
+
+**Note:** Version bump only for package @coveo/atomic-hosted-page
+
+# [0.3.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic-hosted-page@0.2.3...@coveo/atomic-hosted-page@0.3.0) (2023-04-25)
+
+### Features
+
+- **organizationendpoints:** support organization endpoints for hosted-pages components ([#2838](https://github.com/coveo/ui-kit/issues/2838)) ([e0b73f1](https://github.com/coveo/ui-kit/commit/e0b73f16af1a4e794a49621a9e856cad95169cae))
+
 ## [0.2.3](https://github.com/coveo/ui-kit/compare/@coveo/atomic-hosted-page@0.2.2...@coveo/atomic-hosted-page@0.2.3) (2023-04-12)
 
 **Note:** Version bump only for package @coveo/atomic-hosted-page