+++ packages/headless/CHANGELOG.md
@@ -3,6 +3,58 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.16.4](https://github.com/coveo/ui-kit/compare/@coveo/headless@2.16.3...@coveo/headless@2.16.4) (2023-05-11)
+
+**Note:** Version bump only for package @coveo/headless
+
+## [2.16.3](https://github.com/coveo/ui-kit/compare/@coveo/headless@2.16.2...@coveo/headless@2.16.3) (2023-05-11)
+
+**Note:** Version bump only for package @coveo/headless
+
+## [2.16.2](https://github.com/coveo/ui-kit/compare/@coveo/headless@2.16.1...@coveo/headless@2.16.2) (2023-05-09)
+
+### Bug Fixes
+
+- **atomic,headless:** fix support for exact phrase stemming in quickview ([#2860](https://github.com/coveo/ui-kit/issues/2860)) ([a8be9fd](https://github.com/coveo/ui-kit/commit/a8be9fd222bdf187729c55b7b2574f2d0eeb961b))
+
+## [2.16.1](https://github.com/coveo/ui-kit/compare/@coveo/headless@2.16.0...@coveo/headless@2.16.1) (2023-05-08)
+
+**Note:** Version bump only for package @coveo/headless
+
+# [2.16.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@2.15.0...@coveo/headless@2.16.0) (2023-05-04)
+
+### Bug Fixes
+
+- **atomic:** support limited version of quickview when sandbox allow-same-origin not present ([#2866](https://github.com/coveo/ui-kit/issues/2866)) ([83dd211](https://github.com/coveo/ui-kit/commit/83dd211932b325442086360524d5089067bcece5))
+
+### Features
+
+- **headless:** smart snippet questions list controller created for the insight use case ([#2823](https://github.com/coveo/ui-kit/issues/2823)) ([030ff0b](https://github.com/coveo/ui-kit/commit/030ff0b27efcf07f72eac986958562022df73f40))
+
+# [2.15.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@2.14.3...@coveo/headless@2.15.0) (2023-04-27)
+
+### Features
+
+- **smart snippet:** add headless smart snippet support for insight engine ([#2836](https://github.com/coveo/ui-kit/issues/2836)) ([95257b0](https://github.com/coveo/ui-kit/commit/95257b068cceec85ce03855af83a393aeda4350d))
+
+## [2.14.3](https://github.com/coveo/ui-kit/compare/@coveo/headless@2.14.2...@coveo/headless@2.14.3) (2023-04-27)
+
+**Note:** Version bump only for package @coveo/headless
+
+## [2.14.2](https://github.com/coveo/ui-kit/compare/@coveo/headless@2.14.1...@coveo/headless@2.14.2) (2023-04-27)
+
+**Note:** Version bump only for package @coveo/headless
+
+## [2.14.1](https://github.com/coveo/ui-kit/compare/@coveo/headless@2.14.0...@coveo/headless@2.14.1) (2023-04-26)
+
+**Note:** Version bump only for package @coveo/headless
+
+# [2.14.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@2.13.1...@coveo/headless@2.14.0) (2023-04-25)
+
+### Features
+
+- **organizationendpoints:** support organization endpoints for hosted-pages components ([#2838](https://github.com/coveo/ui-kit/issues/2838)) ([e0b73f1](https://github.com/coveo/ui-kit/commit/e0b73f16af1a4e794a49621a9e856cad95169cae))
+
 ## [2.13.1](https://github.com/coveo/ui-kit/compare/@coveo/headless@2.13.0...@coveo/headless@2.13.1) (2023-04-24)
 
 **Note:** Version bump only for package @coveo/headless